requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.4
python-dotenv==1.0.1
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dateutil==2.9.0.post0